+++ bytebank/src/main/kotlin/br/com/alura/bytebank/main.kt
@@ -1,25 +1,35 @@
-import br.com.alura.bytebank.modelo.*
+import br.com.alura.bytebank.modelo.Endereco
 
 fun main() {
 
-    val fran = object : Autenticavel {
-        val nome: String = "Fran"
-        val cpf: String = "111.111.111-11"
-        val senha: Int = 1000
+    println()
+    println(1)
+    println(1.0)
+    println(true)
+    println("teste")
 
-        override fun autentica(senha: Int) = this.senha == senha
-    }
+    val endereco = Endereco()
 
-    val sistemaInterno = SistemaInterno()
-    sistemaInterno.entra(fran, 1000)
+    println(endereco)
 
-    println("Nome do cliente -> ${fran.nome}")
+    imprime()
+    imprime(1)
+    imprime(1.0)
+    imprime(endereco)
+}
 
-    val alex = Cliente(nome = "Alex", cpf = "", senha = 1)
-    val contaPoupanca = ContaPoupanca(titular = alex, numero = 1000)
-    val contaCorrente = ContaCorrente(titular = alex, numero = 1001)
+fun imprime(){
 
-    testaContasDiferentes()
+}
 
-    println("Total de contas: ${Conta.total}")
+fun imprime(valor: Int){
+
+}
+
+fun imprime(valor: Double){
+
+}
+
+fun imprime(endereco: Endereco){
+
 }
